Finish route algorithm and tidyup

+++ src/main/java/groupsix/citywalk/model/BasicTransportMode.java
@@ -8,13 +8,14 @@ import java.util.HashMap;
 public class BasicTransportMode extends TransportMode{
     private Location[] region;
 
-    public BasicTransportMode(String name, int timeFactor, double carbonFactor, Location[] region) {
+    public BasicTransportMode(String name, double timeFactor, int carbonFactor, Location[] region) {
         super(name, timeFactor, carbonFactor);
         this.region = region;
     }
     public Location[] getRegion(){
         return region;
     }
+
     public boolean checkLocation (Location location){
         for (Location loc : region) {
             if (loc.equals(location)) {
